working improv approach for new pybullet env

+++ src/tamp_improv/approaches/improvisational/policies/mpc.py
@@ -43,8 +43,8 @@ class MPCPolicy(Policy[ObsType, ActType]):
         self.action_dims = 1  # Default for Discrete
 
         # Initialize arrays with proper dtypes
-        self._control_times: NDArray[np.float32] = np.zeros(0)
-        self._trajectory_times: NDArray[np.float32] = np.zeros(0)
+        self._control_times: NDArray[np.float32] = np.zeros(0, dtype=np.float32)
+        self._trajectory_times: NDArray[np.float32] = np.zeros(0, dtype=np.float32)
         self.last_solution: NDArray[np.float32] = np.zeros(0, dtype=np.float32)
         self._first_solve = False
 
@@ -291,7 +291,9 @@ class MPCPolicy(Policy[ObsType, ActType]):
 
         trajectories = []
         for pts in new_points:
-            trajectory = np.zeros((self.config.horizon,) + box_space.shape)
+            trajectory = np.zeros(
+                (self.config.horizon,) + box_space.shape, dtype=np.float32
+            )
             for dim in range(pts.shape[-1] if pts.ndim > 1 else 1):
                 idx = (..., dim) if pts.ndim > 1 else ...
                 trajectory[idx] = np.interp(
